+++ src/common/js/startup.js
@@ -3,13 +3,13 @@
  */
 
 Meteor.startup(() => {
-    if( EnvSettings._conf.verbosity & EnvSettings.C.Verbose.STARTUP_MARK ){
+    if( EnvSettings.configure().verbosity & EnvSettings.C.Verbose.STARTUP_MARK ){
         console.log( 'pwix:env-settings startup' );
     }
 });
 
 Meteor.startup(() => {
-    if( EnvSettings._conf.verbosity & EnvSettings.C.Verbose.STARTUP_DUMP ){
+    if( EnvSettings.configure().verbosity & EnvSettings.C.Verbose.STARTUP_DUMP ){
         console.log( 'pwix:env-settings startup', EnvSettings );
     }
 });